move simulated cameras into module
