+++ rosys/vision/simulated_camera/__init__.py
@@ -1,0 +1,7 @@
+from .simulated_camera import SimulatedCamera
+from .simulated_camera_provider import SimulatedCameraProvider
+
+__all__ = [
+    'SimulatedCamera',
+    'SimulatedCameraProvider',
+]
